requirements.txt
echo "flask
gunicorn
requests
Pillow" > requirements.txt